Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main() {
     int x,y,z;
-    int average;
+    float average;
     scanf("%d%d%d",&x,&y,&z);
     average=(x+y+z)/3;
     printf("Average: %.2f",average);
